Renderer MK1 working

+++ Omega-Phi/fvec2.cpp
@@ -36,6 +36,6 @@ bool OP::fvec2::operator!=(const fvec2& other) {
 	return !(*this == other);
 }
 
-int OP::fvec2::dotProduct(const fvec2& other) {
+float OP::fvec2::dotProduct(const fvec2& other) {
 	return (this->x * other.x + this->y * other.y);
 }
